Move Endpoint interface to tfschema package
