+++ internal/tfschema/endpoint.go
@@ -1,4 +1,4 @@
-package provider
+package tfschema
 
 type Endpoint interface {
 	GetMetadata() interface{}
